fix
fix

+++ src/MinimaxResult.java
@@ -1,9 +1,9 @@
 import backstage.move.*;
 public class MinimaxResult {
-    public int score;
+    public double score;
     public Move bestMove;
 
-    public MinimaxResult(int score, Move bestMove) {
+    public MinimaxResult(double score, Move bestMove) {
         this.score = score;
         this.bestMove = bestMove;
     }
